refactor: 입력 담당 클래스 분리
- 문자열 입력, 숫자 입력만을 담당하는 Input클래스 생성
- InputView 안에 있던 입력 문자열, 숫자 입력 기능을 Input 클래스로 매서드 추출
